Fix #6
